# fix(core): fix fetch `size_hint`