--- a/core/ops_builtin.rs
+++ b/core/ops_builtin.rs
@@ -188,10 +188,11 @@ async fn op_read_all(
   let mut grow_len: usize = 64 * 1024;
 
   let (min, maybe_max) = resource.size_hint();
-  // Try to determine an optimal starting buffer size for this resource based
+  // Try to determine an optimial starting buffer size for this resource based
   // on the size hint.
   let initial_size = match (min, maybe_max) {
-    (_min, Some(max)) => max as usize,
+    (min, Some(max)) if min == max => min as usize,
+    (_min, Some(max)) if (max as usize) < grow_len => max as usize,
     (min, _) if (min as usize) < grow_len => grow_len,
     (min, _) => min as usize,
   };