When `max` is known and is greater than `grow_len` initial value is being set to `min` / `grow_len` instead of `max`. This is causing unnecessary resizes.