@marcosc90 2022-10-12
> Max is an upper bound. We don't necessarily want to or can allocate that much space. The idea behind this logic is that we allocate the exact size we need if we know it, and otherwise we start at the lesser of `min` or `grow_length`. This prevents initial over-allocation.
> 
> As such I don't think this change is necessarily beneficial.

If max is known, we should allocate that much space, otherwise, it shouldn't be set. both fetch & http `size_hint` is returning the exact size of the body in `max` (the only two resources setting max currently), which is being ignored completely by `read_all`. What's the use of `max` if we're not allocating the max size at once, currently unnecessary resizes are being done. Setting `max` is specifically for that fast path.

We know exactly the body size, we should use that fast path when available.