@marcosc90 2022-10-13
Furthermore this patch increases `fetch` performance when consuming a body (the most common operation when using `fetch`)

**main**

```
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
runtime: deno 1.26.1 (x86_64-unknown-linux-gnu)

benchmark       time (avg)             (min … max)       p75       p99      p995
-------------------------------------------------- -----------------------------
fetch 1mb   957.65 µs/iter   (538.74 µs … 4.36 ms) 949.99 µs   2.99 ms   3.91 ms
fetch 5mb     4.06 ms/iter     (3.25 ms … 7.66 ms)   4.11 ms   7.51 ms   7.66 ms
fetch 128k  265.63 µs/iter   (162.47 µs … 3.33 ms) 289.15 µs 914.62 µs   1.14 ms
fetch 16kb  127.55 µs/iter (105.26 µs … 953.49 µs) 130.66 µs 255.74 µs 324.11 µs
```

**this patch**

```
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
runtime: deno 1.26.1 (x86_64-unknown-linux-gnu)

benchmark       time (avg)             (min … max)       p75       p99      p995
-------------------------------------------------- -----------------------------
fetch 1mb   613.06 µs/iter   (398.42 µs … 3.29 ms) 715.54 µs   1.42 ms   1.71 ms
fetch 5mb     2.82 ms/iter     (1.69 ms … 6.26 ms)   3.08 ms   5.82 ms   6.26 m
fetch 128k  194.14 µs/iter   (145.34 µs … 1.93 ms) 208.52 µs 454.15 µs 581.11 µs
fetch 16kb  122.34 µs/iter (103.14 µs … 949.53 µs) 119.81 µs 248.41 µs 332.82 µs
```

> This prevents initial over-allocation.

Only two resources overriding `size_hint` set `max` to the exact body size, so it doesn't prevent initial over-allocation, it only causes extra resize operations.


|  |128kb|1mb|5mb|2gb
|--|--|--|--|--|
| main: resizes |1|6|14|4100
| patch: resizes |0|0|0|0|0