@marcosc90 2022-10-19
> Sorry for not responding to this earlier. The logic in `op_read_all` is correct. The size_hint for the HTTP resources is just not. It currently is:
> 
> ```
>   fn size_hint(&self) -> (u64, Option<u64>) {
>     (0, self.size)
>   }
> ```
> 
> but it should be
> 
> ```
>   fn size_hint(&self) -> (u64, Option<u64>) {
>     (self.size.unwrap_or(0), self.size)
>   }
> ```
> 
> If the **exact** size is known (`self.size == Some(_)`), then the upper and lower bound should both be set to that exact size. Once you do this the existing `op_read_all` optimisation starts working.
> 
> This is already the case for the `size_hint()` in `ext/http`.

While that will fix the performance issue, IMO that's patching incorrect logic of `op_read_all`. There's no need to set `max` at all if we're using `min` instead and ignoring `max` when it's set. I'll submit the changes for fetch `size_hint` to improve performance though.