@lucacasonato 2022-10-19
We're not ignoring max. We're just ignoring max for pre-allocation. But we do use max to prevent over-allocation for example. Under-allocation is always better than over-allocation, as over-allocation can hard crash the process with an OOM error.